+++ Game/knight.py
@@ -1,5 +1,0 @@
-from Game.pieces import Piece
-
-class Knight(Piece):
-    white_str = "♘"
-    black_str = "♞"