+++ bishop.py
@@ -1,4 +1,4 @@
-from Game.pieces import Piece
+from pieces import Piece
 
 class Bishop(Piece):
     white_str = "♗"