+++ chess.py
@@ -1,5 +1,5 @@
-from Game.board import *
-from Game.exceptions import *
+from board import *
+from exceptions import *
 
 class Chess:
     def __init__(self):